SDC-5341: JDK7: Switch build to use JDK8
Reviewed-on: https://review.streamsets.net/6659
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>

+++ container/src/test/java/com/streamsets/datacollector/main/TestRuntimeInfo.java
@@ -93,7 +93,7 @@ public class TestRuntimeInfo {
     Assert.assertFalse(info.hasAttribute("a"));
     info.setAttribute("a", 1);
     Assert.assertTrue(info.hasAttribute("a"));
-    Assert.assertEquals(1, info.getAttribute("a"));
+    Assert.assertEquals(1, (int)info.getAttribute("a"));
     info.removeAttribute("a");
     Assert.assertFalse(info.hasAttribute("a"));
   }
